requirements.txt
aiogram==3.10.0
pymongo==4.8.0
python-dotenv==1.0.1
Requests==2.32.3